fix: track USDT in StableTreasury instead of CUSD, move reward vaults to treasury adapter
Addresses RohanNero review:
1. Track real USDT balances in StableTreasury (backs all minted CUSD)
2. Remove CUSD-to-USDT misrepresentation - track the actual backing asset
3. Move protocol-owned reward vaults from staking to projects/treasury/comdex.js
