+++ projects/comdex/index.js
@@ -1,68 +1,43 @@
 const ADDRESSES = require('../helper/coreAssets.json');
+const { sumTokens2 } = require('../helper/unwrapLPs');
 
-// CUSD Token (6 decimals) — settlement currency (1:1 USD peg, not listed on CoinGecko)
-const CUSD = "0xc5079966b3190909f69306fE7587ffE493dEdB5F";
-
-// BSC USDT (18 decimals) — used as pricing proxy for CUSD
+// BSC USDT (18 decimals) — the real backing asset
 const BSC_USDT = ADDRESSES.bsc.USDT;
 
-// All Treasury contracts that hold CUSD backing tokenized assets
+// CUSD Token (6 decimals) — protocol's internal settlement stablecoin
+// Minted 1:1 when users deposit USDT via ComdexStableMarket.
+// Source code: contracts/src/tokens/CUSDToken.sol
+const CUSD = "0xc5079966b3190909f69306fE7587ffE493dEdB5F";
+
+// ── Treasury contracts ──
+// StableTreasury holds USDT backing (received when users buy CUSD).
+// Commodity Treasuries hold CUSD (received when users buy commodity tokens).
 const TREASURIES = [
+  "0xD8875eEf762A6C23f8473E19C896B584BAaF007A",   // Stable Treasury (holds USDT)
   "0x2581b28e9f261c0ab5533dbf4305a806afb2fe1e",   // Gold Treasury
   "0x0893e45ad6e655787be1e669e54d9c237b1ff083",   // Silver Treasury
   "0x216238dce287b8b8f2eda8842040c8f862c776cb",   // Platinum Treasury
   "0xc9298ef68392c48f521f7cb8c8261c88099c4b36",   // Palladium Treasury
   "0x5aeAA55d5024CEf2c32497be59b7506481fCddbD",   // Oil Treasury
   "0x9621bD2eF645cb41356Da9Cab8d9DB1EC4e3be1A",   // Copper Treasury
-  "0xD8875eEf762A6C23f8473E19C896B584BAaF007A",   // Stable Treasury (V3)
+  "0x7C2358a7a67478036A5287022D81EAa760E644B1",   // Aluminum Treasury
+  "0xD1c8dcF50FBA73FE119cC284454cF8fC3ecdd3Fe",   // Zinc Treasury
+  "0x1beFfff48a7d33b6be43246B9F50147845f69e98",   // Nickel Treasury
+  "0xd04b370a88Dc16840CdefB53051168Eb983d113a",   // Lead Treasury
 ];
-
-// Protocol-owned CUSD reserves earmarked for staking incentive distribution.
-// These are NOT user-staked positions — they are protocol inventory.
-const REWARD_VAULTS = [
-  "0x9057c3a25ff8e71bc05782a3d44a74fa7eb95688",   // Gold Reward Vault
-  "0x54ecbcb04da981f3a6896ad1b83c5ec47ee4d618",   // Silver Reward Vault
-  "0xf98f5908fa0b2b0cb32b79f2612446c7e3bffcad",   // Platinum Reward Vault
-  "0x2f30fd7b18b8c69e85131e387d88919ae85f26c1",   // Palladium Reward Vault
-];
-
-/**
- * Read CUSD balanceOf for a list of owners, convert to USDT-equivalent.
- * CUSD is 6 decimals, BSC USDT is 18 decimals → multiply by 1e12.
- */
-async function getCusdBalances(api, owners) {
-  const bals = await api.multiCall({
-    abi: 'erc20:balanceOf',
-    calls: owners.map(owner => ({ target: CUSD, params: [owner] })),
-  });
-  let total = BigInt(0);
-  for (let i = 0; i < bals.length; i++) {
-    const bal = bals[i];
-    if (bal == null || bal === '') {
-      throw new Error(`balanceOf returned null/empty for CUSD at ${owners[i]}`);
-    }
-    total += BigInt(bal);
-  }
-  // Scale from 6 decimals (CUSD) to 18 decimals (USDT)
-  api.add(BSC_USDT, (total * BigInt(1e12)).toString());
-}
 
 async function tvl(api) {
-  // Sum CUSD in all Treasury contracts → report as USDT
-  await getCusdBalances(api, TREASURIES);
-}
-
-async function staking(api) {
-  // Protocol-owned CUSD reserves funding staking rewards.
-  // These are protocol inventory, not user-staked positions.
-  await getCusdBalances(api, REWARD_VAULTS);
+  const tokensAndOwners = [];
+  for (const treasury of TREASURIES) {
+    tokensAndOwners.push([BSC_USDT, treasury]);  // USDT balances in all treasuries
+    tokensAndOwners.push([CUSD, treasury]);       // CUSD balances in all treasuries
+  }
+  return sumTokens2({ api, tokensAndOwners });
 }
 
 module.exports = {
-  misrepresentedTokens: true,
-  methodology: "TVL is the sum of CUSD (USD-pegged stablecoin) held in Treasury contracts backing tokenized commodity tokens. CUSD balances are reported as USDT equivalent (misrepresentedTokens). Protocol-owned Reward Vault reserves (CUSD earmarked for staking incentive distribution) are tracked under 'staking' as the closest available category — these are not user-staked positions. Staked commodity tokens are excluded to avoid double-counting with Treasury reserves.",
+  methodology: "TVL is the sum of USDT and CUSD held across all Treasury contracts. The StableTreasury holds USDT (deposited when users mint CUSD 1:1). Commodity Treasuries (Gold, Silver, Platinum, Palladium, Oil, Copper, Aluminum, Zinc, Nickel, Lead) hold CUSD received from commodity token purchases.",
   bsc: {
     tvl,
-    staking,
   },
 };
